UI Beta
All the interfaces we need (to be revised).
As soon we will have a db or the saving/ recall .text, will be possible
to implement the rest.

The look and fell is crap but it works.
Each time an action is performed, the old JFrame is closed and a new
one is opened.